fix: quote 3.0 in CI configuration
An unquoted 3.0 is truncated to a 3, and loads the latest Ruby 3 version - at this time Ruby 3.1.2.

There is a commented out "3.1", so I'm not sure if this behavior is intentional.  If there's a desired to run exclusively with the latest Ruby 3 version, I'd suggest that the ambiguous 3.0 be replaced with a 3.  Otherwise it might be desirable to uncomment the 3.1, since it runs fine.