Map citext type from PostgreSQL to string type in TypeScript
`citext` is case-insensitive character string type in PostgreSQL. It's
often used for unique fields where case sensitivity should be ignored,
like email, repo name, etc. Supporting this mapping by default would
make it easier to integrate this library in a codebase that is already
using `citext`. This mapping is straightforward and it shouldn't require
developers to extend the gem configuration to support it.