Remove pry-byebug gem
The `pry-byebug` gem dependency is causing an exception when running the
test suit. This is most likely due to new Ruby version. Unfortunately,
there is no recent updates in the gem and it doesn't seem like a fix
would be applied there.

The dependency is not used directly and my guess is that it's there
mainly for debugging purposes. I would propose to drop this dependency
and use the built-in `binding.irb` for debugging.

https://ruby-doc.org/stdlib-2.7.0/libdoc/irb/rdoc/Binding.html

Here's the output from trying to run the test suite.

Before:

```
❯ bin/rspec

An error occurred while loading spec_helper.
Failure/Error: require "pry-byebug"

NameError:
  undefined method `=~' for class `Pry::Code'
No examples found.

Finished in 0.00002 seconds (files took 0.24819 seconds to load)
0 examples, 0 failures, 1 error occurred outside of examples
```

After:

```
❯ bin/rspec
.Generating TypeScript interfaces...completed in 0.0 seconds.
Found 9 serializers:
        ComposerSerializer
        ComposerWithSongsSerializer
        ModelSerializer
        Nested::AlbumSerializer
        SnakeComposerSerializer
        SongSerializer
        SongWithVideosSerializer
        VideoSerializer
        VideoWithSongSerializer
.

Finished in 0.15912 seconds (files took 1.01 seconds to load)
2 examples, 0 failures
```